requirements.txt
tensorflow-gpu~=2.9.1
Pillow~=9.1.1
numpy~=1.22.4
matplotlib~=3.5.2
pandas~=1.4.2
scipy~=1.8.1
